feat: add LEGORN env vars to AppCenter
We want to add environment variables prefixed with LEGORNS? to AppCenter build configuration, this
allows us to hide secrets in GitHub and inject them at build time using RNConfig. The reason we want
this in a script rather than in AppCenter directly, is that we cannot specify secrets for multiple
branches, i.e. release/x.y.z, thus we'd need to set up secrets each time we create a new release
branch.